+++ src/pages/User/LostList/LostList.tsx
@@ -1,22 +1,14 @@
-import { useState, ChangeEvent } from 'react';
-import { useSearchParams } from 'react-router-dom';
 import LostList from '../../../components/Lost/Lostlist';
-import Input from '../../../shared/ui/Input/Input';
+
 import TopBar from '../../../shared/ui/TopBar/TopBar';
 import { useNavigate } from 'react-router-dom';
+import LostInput from '../../../components/Lost/LostInput';
+import { useState } from 'react';
 
 const LostListPage = () => {
   const navigate = useNavigate();
+  const [mq, setMq] = useState('');
 
-  const [searchValue, setSearchValue] = useState('');
-
-  const onChangeSearchValue = (event: ChangeEvent<HTMLInputElement>) => {
-    setSearchValue(event.target.value);
-  };
-  const [searchParams] = useSearchParams({
-    mq: '',
-  });
-  const mq = searchParams.get('mq');
   return (
     <div className="flex flex-col justify-center items-center ">
       <TopBar
@@ -25,12 +17,7 @@ const LostListPage = () => {
         onSecondClick={() => navigate('/user/lost-list')}
       />
       <div className="mt-[25px]">
-        <Input
-          placeholder="검색어를 입력해주세요"
-          title="검색"
-          value={searchValue}
-          onChange={onChangeSearchValue}
-        />
+        <LostInput setSearchValue={setMq} />
         <LostList searchValue={mq} />
       </div>
     </div>
